+++ CHANGELOG.md
@@ -12,6 +12,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
   `ManagedPreparedStatement`. The behavior of `CassandraResultSet.close()` method has been adapted in consequence.
 - Implement methods `absolute(int)`, `afterLast()`, `beforeFirst()`, `first()`, `last()`, `previous()` and 
   `relative(int)` in `CassandraResultSet`.
+- Implement some JDBC API 4.3 methods:
+  - in `CassandraDatabaseMetaData`: `getMaxLogicalLobSize()`, `supportsRefCursors()` and `supportsSharding()`.
+  - in `CassandraDataSource`: `createConnectionBuilder()`.
+  - in `Statement` implementations: methods `enquoteIdentifier(String, boolean)` and `isSimpleIdentifier(String)` are
+    adapted to Cassandra-specific rules.
 ### Changed
 - **BREAKING CHANGE**: upgrade minimal required JDK to 17.
 - Update supported hosts for Amazon Keyspaces: add `me-central-1` and dual-stack endpoints (`*.api.aws`).